requirements.txt
aiokafka==0.10.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
confluent-kafka==2.3.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
kafka-python==2.0.2
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
websockets==12.0
